WOW
You really demonstrated a terrific understanding of the language and your ability to apply the principles to real challenges to achieve a goal. Great work! This project is beyond what you were required to do, and I really hope you're moving foward with code and development. The only thing I would suggest moving forward to improve upon would be to comment your code more, so that users who are new to the codebase can get a grasp on what you;re doing. Would you mind if I use this as an example in class?